fixed mesh STL persistence typo
git-svn-id: http://b035a0a.online-server.cloud/svn/ITDevelopment@1106 21205d9f-8918-2240-a3d5-c39064a2c852

+++ Include/imt3d/CMesh3d.cpp
@@ -184,7 +184,7 @@ bool CMesh3d::SaveToStlFile(const QString& filePath) const
 
 	// write triangles count
 	quint32 trianglesCount = static_cast<quint32>(m_indices.size());
-	bool retVal = WriteTypedValue<quint16>(trianglesCount, file);
+	bool retVal = WriteTypedValue<quint32>(trianglesCount, file);
 
 	// write triangles
 	for (size_t i = 0; retVal && i < m_indices.size(); i++){
